+++ engine/src/test/scala/org/scalarules/dsl/nl/grammar/datum/DatumTest.scala
@@ -1,9 +1,9 @@
-package org.scalarules.dsl.nl.grammar.date
+package org.scalarules.dsl.nl.grammar.datum
 
 import org.joda.time.LocalDate
 import org.scalarules.dsl.nl.grammar.aanwezig
 import org.scalarules.utils.InternalBerekeningenTester
-import org.scalarules.dsl.nl.grammar.date.DatumTestGlossary._
+import org.scalarules.dsl.nl.grammar.datum.DatumImplicits
 
 class DatumTest extends InternalBerekeningenTester(new DatumTestsBerekening) with DatumImplicits {
