+++ engine/src/test/scala/org/scalarules/dsl/nl/grammar/datum/DatumTestsBerekening.scala
@@ -1,7 +1,6 @@
-package org.scalarules.dsl.nl.grammar.date
+package org.scalarules.dsl.nl.grammar.datum
 
 import org.scalarules.dsl.nl.grammar._
-import org.scalarules.dsl.nl.grammar.date.DatumTestGlossary._
 
 class DatumTestsBerekening extends Berekening (
